// src/types/playerDetail.ts
export type PlayerDetail = {
  id?: string;
  name: string;
  avatar?: string;
  age?: string;
  nationality?: string;
  position?: string;
  club?: string;
  jersey?: string;
  height?: string;
  weight?: string;
  foot?: string;
  form?: string;
  marketValue?: string;
  contractUntil?: string;
  salary?: string;
};